Contact and User classes
User class is the main class. This class handles the actual subscriber/user of the service. It holds their name, age, date of birth, username, password, and list of contacts.

Contact class is a class that holds the list of contacts from the user. Each contact has a name, email, and phone number.